+++ clients/utils/src/components.rs
@@ -7,6 +7,7 @@ use ekiden_registry_client;
 use ekiden_roothash_client;
 use ekiden_scheduler_client;
 use ekiden_storage_frontend;
+use ekiden_storage_multilayer;
 
 /// Register known components for dependency injection.
 pub fn register_components(known_components: &mut KnownComponents) {
@@ -27,6 +28,7 @@ pub fn register_components(known_components: &mut KnownComponents) {
     ekiden_roothash_client::RootHashClient::register(known_components);
     // Storage.
     ekiden_storage_frontend::StorageClient::register(known_components);
+    ekiden_storage_multilayer::MultilayerBackend::register(known_components);
 }
 
 /// Create known component registry.